+++ web/app/scripts/compiled-es5/directives/form.js
@@ -59,7 +59,10 @@ _module.directive('surveysForm', [_api2.default.componentName, _fieldsetsParser2
 
             scope.submit = function () {
                 if (scope.responseUrl) {
-                    API.post(scope.responseUrl, scope.model).then(scope.onSuccess).catch(function (response) {
+                    API.post(scope.responseUrl, scope.model).then(function () {
+                        FieldsetParser.addFieldErrors(scope.fields, {});
+                        scope.onSuccess();
+                    }).catch(function (response) {
                         var errors = response && response.data;
                         FieldsetParser.addFieldErrors(scope.fields, errors);
                         scope.onFailure();