+++ web/app/scripts/compiled-es5/services/fields-config.js
@@ -46,6 +46,11 @@ _module.run(['formlyConfig', moduleProperties.componentName, function (formlyCon
         templateUrl: templatesBase + '/radio.html'
     });
 
+    formlyConfig.setType({
+        name: 'proportion',
+        templateUrl: templatesBase + '/proportion.html'
+    });
+
     formlyConfig.setWrapper([{
         name: 'panel',
         templateUrl: templatesBase + '/wrapper.html'