+++ web/app/scripts/src/main.es.js
@@ -4,6 +4,7 @@ import API from 'services/api';
 import FieldsConfig from 'services/fields-config';
 import SurveysFormDirective from 'directives/form';
 import CalculatePercentageDirective from 'directives/calculate-percentage';
+import ProportionFieldDirective from 'directives/proportion-field';
 
 angular.module('incuna-surveys', [
     'drf-form-field',
@@ -12,5 +13,6 @@ angular.module('incuna-surveys', [
     FieldsConfig.moduleName,
     API.moduleName,
     SurveysFormDirective.moduleName,
-    CalculatePercentageDirective.moduleName
+    CalculatePercentageDirective.moduleName,
+    ProportionFieldDirective.moduleName
 ]);