+++ web/tests/survey-form-spec.js
@@ -115,6 +115,10 @@ describe('surveysForm directive', function () {
 
         describe('success callback', function () {
             beforeEach(function () {
+                const responseDefer = this.$q.defer();
+                responseDefer.resolve(this.formResponse);
+                spyOn(this.API, 'post').and.returnValue(responseDefer.promise);
+                spyOn(this.FieldsParser, 'addFieldErrors');
             });
 
             it('should be defiend', function () {
@@ -122,13 +126,16 @@ describe('surveysForm directive', function () {
             });
 
             it('should be called if the post succeeds', function () {
-                const responseDefer = this.$q.defer();
-                responseDefer.resolve(this.formResponse);
-                spyOn(this.API, 'post').and.returnValue(responseDefer.promise);
                 this.isolated.submit();
                 expect(this.API.post).toHaveBeenCalledWith(responseUrl, this.formResponse)
                 this.scope.$digest();
                 expect(this.scope.mySubmit).toHaveBeenCalledWith();
+            });
+
+            it('should call addFieldErrors if the post succeeds', function () {
+                this.isolated.submit();
+                this.scope.$digest();
+                expect(this.FieldsParser.addFieldErrors).toHaveBeenCalledWith(fields, {});
             });
         });
 
@@ -149,7 +156,6 @@ describe('surveysForm directive', function () {
             it('should call addFieldErrors if the post fails', function () {
                 this.isolated.submit();
                 this.scope.$digest();
-
                 expect(this.FieldsParser.addFieldErrors).toHaveBeenCalledWith(fields, postErrors);
             });